requirements.backup.txt
aider-chat==0.46.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
defusedxml==0.7.1
diff-match-patch==20230430
diskcache==5.6.3
distro==1.9.0
et_xmlfile==2.0.0
filelock==3.15.4
flake8==7.1.0
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.108.0
google-auth==2.38.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grep-ast==0.3.2
grpcio==1.72.0rc1
grpcio-status==1.71.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.0
idna==3.7
importlib_metadata==7.2.1
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
litellm==1.41.27
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.37.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==10.4.0
prompt_toolkit==3.0.47
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.0
pycparser==2.22
pydantic>=2.5.3
pydantic_core==2.14.6
pyflakes==3.2.0
pygame==2.6.1
Pygments==2.18.0
pypandoc==1.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
scipy==1.13.1
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.7
soundfile==0.12.1
soupsieve==2.5
tenacity==9.1.2
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.4
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.2
uv==0.6.14
wcwidth==0.2.13
Werkzeug==3.1.3
windows-curses==2.3.3 ; sys_platform == 'win32'
yarl==1.9.4
youtube-transcript-api==1.0.3
zipp==3.19.2
